68. useReducer - Remove Item complete
