+++ tutorials/react_advanced/src/tutorial/6-useReducer/setup/Modal.js
@@ -1,6 +1,11 @@
 import React, { useEffect } from "react";
 
-const Modal = ({ modalContent }) => {
+const Modal = ({ modalContent, closeModal }) => {
+	useEffect(() => {
+		setTimeout(() => {
+			closeModal();
+		}, 3000);
+	});
 	return (
 		<div className="modal">
 			<p>{modalContent}</p>
