+++ tutorials/react_advanced/src/tutorial/6-useReducer/setup/index.js
@@ -2,26 +2,8 @@ import React, { useState, useReducer } from "react";
 import Modal from "./Modal";
 import { data } from "../../../data";
 // reducer function
-const reducer = (state, action) => {
-	console.log(state);
-	if (action.type === "ADD_ITEM") {
-		const newPeople = [...state.people, action.payload];
-		return {
-			...state,
-			people: newPeople,
-			isModalOpen: true,
-			modalContent: "item added",
-		};
-	}
-	if (action.type === "NO_VALUE") {
-		return {
-			...state,
-			isModalOpen: true,
-			modalContent: "please enter value",
-		};
-	}
-	throw new Error("no matching action type");
-};
+import { reducer } from "./reducer";
+
 const defaultState = {
 	people: [],
 	isModalOpen: false,
@@ -40,10 +22,17 @@ const Index = () => {
 			dispatch({ type: "NO_VALUE" });
 		}
 	};
-
+	const closeModal = () => {
+		dispatch({ type: "CLOSE_MODAL" });
+	};
 	return (
 		<>
-			{state.isModalOpen && <Modal modalContent={state.modalContent} />}
+			{state.isModalOpen && (
+				<Modal
+					closeModal={closeModal}
+					modalContent={state.modalContent}
+				/>
+			)}
 			<form onSubmit={handleSubmit}>
 				<div>
 					<input
@@ -56,8 +45,18 @@ const Index = () => {
 			</form>
 			{state.people.map((person) => {
 				return (
-					<div key={person.id}>
+					<div key={person.id} className="item">
 						<h4>{person.name}</h4>
+						<button
+							onClick={() =>
+								dispatch({
+									type: "REMOVE_ITEM",
+									payload: person.id,
+								})
+							}
+						>
+							remove
+						</button>
 					</div>
 				);
 			})}
